requirements.txt
langchain
wikipedia
langchain-community
langchain-openai
langchain-anthropic
langchain_google_genai
python-dotenv
pydantic
duckduckgo-search